requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
Babel==2.8.0
billiard==3.6.2.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect2==0.2.2
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.3
django-celery-results==1.2.0
django-environ==0.4.5
django-extensions==2.2.8
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.2
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isodate==0.6.0
kombu==4.6.7
msgpack==0.6.2
Pillow==7.0.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
rdflib==4.2.2
redis==3.4.1
requests==2.23.0
service-identity==18.1.0
six==1.14.0
SPARQLWrapper==1.8.5
sqlparse==0.3.0
text-unidecode==1.3
tinycss2==1.0.2
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
Wikidata==0.6.1
zope.interface==4.7.2
